various updates

+++ recipes/install.rb
@@ -38,7 +38,8 @@ dpkg_package node[:memsql][:version] do
 end
 
 #move to mounted ec2 volume
-if !Dir.exists?('/data/memsql')
+if !Dir.exists?('/data/memsql') and Dir.exists?('/var/lib/memsql')
+  %x(mkdir -p /data)
   %x(mv /var/lib/memsql /data/memsql)
   %x(ln -s /data/memsql /var/lib/memsql)
 end
@@ -76,6 +77,7 @@ template "/var/lib/memsql/memsql.cnf" do
                 :is_master => node.run_list.roles.include?("master_aggregator") ? true : false
             })
 end
+
 
 node[:memsql][:users].each do |user|
   %x(sudo mysql -u root -h #{master_aggregator[:ipaddress]} -e "grant all on *.* to '#{user[:name]}'@'localhost' identified by '#{user[:password]}'; flush privileges;")
